TINY-12007: Moved acid color picker into silver
