--- modules/acid/src/demo/i18n/en.ts
@@ -1,21 +1,0 @@
-const strings: Record<string, string> = {
-  // 'colorcustom.palette.thumb': 'Palette Picker',
-  // 'colorcustom.palette.spectrum': 'Saturation and Brightness Palette',
-  // 'colorcustom.hue.thumb':'Hue Slider',
-  // 'colorcustom.hue.spectrum': 'Hue Rail',
-  'aria.color.picker': 'Color Picker',
-  'aria.input.invalid': 'Invalid value',
-  'colorcustom.rgb.red.label': 'R',
-  'colorcustom.rgb.red.description': 'Red component',
-  'colorcustom.rgb.green.label': 'G',
-  'colorcustom.rgb.green.description': 'Green component',
-  'colorcustom.rgb.blue.label': 'B',
-  'colorcustom.rgb.blue.description': 'Blue component',
-  'colorcustom.rgb.hex.label': '#',
-  'colorcustom.rgb.hex.description': 'Hex colour code',
-  'colorcustom.rgb.range': 'Range 0 to 255'
-};
-
-export {
-  strings
-};
