+++ modules/acid/src/main/ts/ephox/acid/api/Main.ts
@@ -1,13 +1,13 @@
-import * as ColourPicker from '../gui/ColourPicker';
+import * as ColourTypes from './colour/ColourTypes';
 import * as HexColour from './colour/HexColour';
 import * as HsvColour from './colour/HsvColour';
 import * as RgbaColour from './colour/RgbaColour';
 import * as Transformations from './colour/Transformations';
 
 export {
+  ColourTypes,
   HexColour,
   HsvColour,
   RgbaColour,
-  ColourPicker,
   Transformations
 };
