+++ modules/acid/tsconfig.json
@@ -7,20 +7,14 @@
     "tsBuildInfoFile": "lib/tsconfig.tsbuildinfo",
     "paths": {
       "ephox/acid/test/*": ["src/test/ts/module/ephox/acid/test/*"],
-      "ephox/acid/demo/*": ["src/demo/ts/ephox/acid/demo/*"],
       "ephox/acid/*": ["src/main/ts/ephox/acid/*"]
     }
   },
   "include": [
-    "src/demo/ts",
-    "src/demo/i18n",
     "src/main/ts",
     "src/test/ts"
   ],
   "references": [
-    { "path": "../alloy" },
-    { "path": "../boulder" },
-    { "path": "../katamari" },
-    { "path": "../sugar" }
+    { "path": "../katamari" }
   ]
 }
