+++ modules/tinymce/src/themes/silver/main/ts/ui/colorpicker/I18n.ts
