+++ modules/tinymce/src/themes/silver/main/ts/ui/colorpicker/components/HueSlider.ts
